clean fmt fix mockRepo for tests

+++ internal/app/handlers/mock_RepositoryInterface.go
@@ -15,11 +15,11 @@ type MockRepositoryInterface struct {
 }
 
 // ChangeOrderStatus provides a mock function with given fields: ctx, order, status, accrual
-func (_m *MockRepositoryInterface) ChangeOrderStatus(ctx context.Context, order string, status string, accrual int) error {
+func (_m *MockRepositoryInterface) ChangeOrderStatus(ctx context.Context, order string, status string, accrual float64) error {
 	ret := _m.Called(ctx, order, status, accrual)
 
 	var r0 error
-	if rf, ok := ret.Get(0).(func(context.Context, string, string, int) error); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, string, float64) error); ok {
 		r0 = rf(ctx, order, status, accrual)
 	} else {
 		r0 = ret.Error(0)
